LoginBox 삭제 후 LoginPage에 통합, 라우팅 변경
